Suppress mode when keyboard macros are executing
This fixes #5 by disabling the scroll code whenever we detect that a
keyboard macro is currently executing. This is to address the underlying
cause of early termination of macros as the `scroll-scroll-*` functions
can yield an error by attempting to scroll when we cannot.

NOTE: If the throwing of "End of Buffer" causes other issues we should
      remove this temporary fix and address the root cause of the error.